requirements.txt
langchain < 0.99
langchain-openai
openai < 1.16.99
streamlit < 1.33.99
streamlit-folium < 0.19.99
jupyter
jupyterlab
pandas
python-dotenv